Refactor environment to separate it from pyevm
Preparation to handle #104
