+++ boa/network.py
@@ -9,7 +9,7 @@ from math import ceil
 from eth_account import Account
 from requests.exceptions import HTTPError
 
-from boa.environment import Address, Env
+from boa.environment import Env
 from boa.rpc import (
     EthereumRPC,
     RPCError,
@@ -19,6 +19,7 @@ from boa.rpc import (
     to_int,
     trim_dict,
 )
+from boa.util.abi import Address
 
 
 class TraceObject:
@@ -119,7 +120,7 @@ class NetworkEnv(Env):
         if not self._rpc_has_snapshot:
             raise RuntimeError("RPC does not have `evm_snapshot` capability!")
         try:
-            blkid = self.vm.state._account_db._block_id
+            block_id = self.evm.block_id
             snapshot_id = self._rpc.fetch("evm_snapshot", [])
             yield
             # note we cannot call super.anchor() because vm/accountdb fork
@@ -127,7 +128,7 @@ class NetworkEnv(Env):
         finally:
             self._rpc.fetch("evm_revert", [snapshot_id])
             # wipe forked state
-            self._reset_fork(blkid)
+            self._reset_fork(block_id)
 
     # add account, or "Account-like" object. MUST expose
     # `sign_transaction` or `send_transaction` method!
@@ -183,10 +184,10 @@ class NetworkEnv(Env):
         # non eip-1559 transaction
         return self._rpc.fetch_multi([("eth_gasPrice", []), ("eth_chainId", [])])
 
-    def _check_sender(self, address):
+    def _check_sender(self, address: Address):
         if address is None:
             raise ValueError("No sender!")
-        return Address(address)
+        return address
 
     # OVERRIDES
     def execute_code(
