+++ boa/vm/py_evm.py
@@ -1,0 +1,545 @@
+"""
+A wrapper for the py-evm engine to provide a consistent interface for the environment.
+Handles low level details around state and py-evm tracing.
+"""
+
+import contextlib
+import logging
+import sys
+import warnings
+from typing import Any, Iterator, Optional
+
+import eth.constants as constants
+import eth.tools.builder.chain as chain
+import eth.vm.forks.spurious_dragon.computation as spurious_dragon
+from eth._utils.address import generate_contract_address
+from eth.chains.mainnet import MainnetChain
+from eth.db.atomic import AtomicDB
+from eth.exceptions import Halt
+from eth.typing import JournalDBCheckpoint
+from eth.vm.code_stream import CodeStream
+from eth.vm.message import Message
+from eth.vm.opcode_values import STOP
+from eth.vm.transaction_context import BaseTransactionContext
+from eth_typing import Hash32
+from eth_utils import setup_DEBUG2_logging
+
+from boa.util.abi import Address, abi_decode
+from boa.util.eip1167 import extract_eip1167_address, is_eip1167_contract
+from boa.vm.fast_accountdb import patch_pyevm_state_object, unpatch_pyevm_state_object
+from boa.vm.fork import AccountDBFork
+from boa.vm.gas_meters import GasMeter
+from boa.vm.utils import to_bytes, to_int
+
+
+def enable_pyevm_verbose_logging():
+    logging.basicConfig()
+    logger = logging.getLogger("eth.vm.computation.Computation")
+    setup_DEBUG2_logging()
+    logger.setLevel("DEBUG2")
+
+
+class VMPatcher:
+    _exc_patchables = {
+        # env vars vyper supports
+        "block_number": "_block_number",
+        "timestamp": "_timestamp",
+        "coinbase": "_coinbase",
+        "difficulty": "_difficulty",
+        "prev_hashes": "_prev_hashes",
+        "chain_id": "_chain_id",
+        "gas_limit": "_gas_limit",
+    }
+
+    _cmp_patchables = {"code_size_limit": "EIP170_CODE_SIZE_LIMIT"}
+
+    def __init__(self, vm):
+        patchables = [
+            (self._exc_patchables, vm.state.execution_context),
+            (self._cmp_patchables, spurious_dragon),
+        ]
+        # https://stackoverflow.com/a/12999019
+        object.__setattr__(self, "_patchables", patchables)
+
+    def __getattr__(self, attr):
+        for s, p in self._patchables:
+            if attr in s:
+                return getattr(p, s[attr])
+        raise AttributeError(attr)
+
+    def __setattr__(self, attr, value):
+        for s, p in self._patchables:
+            if attr in s:
+                setattr(p, s[attr], value)
+                return
+
+    # to help auto-complete
+    def __dir__(self):
+        patchable_keys = [k for p, _ in self._patchables for k in p]
+        return dir(super()) + patchable_keys
+
+    # save and restore patch values
+    @contextlib.contextmanager
+    def anchor(self):
+        snap = {}
+        for s, _ in self._patchables:
+            for attr in s:
+                snap[attr] = getattr(self, attr)
+
+        try:
+            yield
+
+        finally:
+            for s, _ in self._patchables:
+                for attr in s:
+                    setattr(self, attr, snap[attr])
+
+
+_opcode_overrides = {}
+
+
+def patch_opcode(opcode_value, fn):
+    global _opcode_overrides
+    _opcode_overrides[opcode_value] = fn
+
+
+# _precompiles is a global which is loaded to the env computation
+# every time one is created. the reasoning is that it would seem
+# confusing to have registered precompiles not persist across envs -
+# if somebody registers a precompile, presumably they want it to work
+# on all envs.
+_precompiles = {}
+
+
+def register_precompile(*args, **kwargs):
+    warnings.warn(
+        "register_recompile has been renamed to register_raw_precompile!", stacklevel=2
+    )
+
+
+def register_raw_precompile(address, fn, force=False):
+    global _precompiles
+    address = Address(address)
+    if address in _precompiles and not force:
+        raise ValueError(f"Already registered: {address}")
+    _precompiles[address.canonical_address] = fn
+
+
+def deregister_raw_precompile(address, force=True):
+    address = Address(address).canonical_address
+    if address not in _precompiles and not force:
+        raise ValueError("Not registered: {address}")
+    _precompiles.pop(address, None)
+
+
+def console_log(computation):
+    msgdata = computation.msg.data_as_bytes
+    schema, payload = abi_decode("(string,bytes)", msgdata[4:])
+    data = abi_decode(schema, payload)
+    print(*data, file=sys.stderr)
+    return computation
+
+
+CONSOLE_ADDRESS = bytes.fromhex("000000000000000000636F6E736F6C652E6C6F67")
+
+register_raw_precompile(CONSOLE_ADDRESS, console_log)
+
+
+# a code stream which keeps a trace of opcodes it has executed
+class TracingCodeStream(CodeStream):
+    __slots__ = [
+        "_length_cache",
+        "_fake_codesize",
+        "_raw_code_bytes",
+        "invalid_positions",
+        "valid_positions",
+        "program_counter",
+    ]
+
+    def __init__(self, *args, start_pc=0, fake_codesize=None, contract=None, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._trace = []  # trace of opcodes that were run
+        self.program_counter = start_pc  # configurable start PC
+        self._fake_codesize = fake_codesize  # what CODESIZE returns
+
+    def __iter__(self) -> Iterator[int]:
+        # upstream says: "a very performance-sensitive method"
+        # note: not clear to me that len(raw_code_bytes) is a hotspot
+        while self.program_counter < self._length_cache:
+            opcode = self._raw_code_bytes[self.program_counter]
+
+            self._trace.append(self.program_counter)
+            self.program_counter += 1
+            yield opcode
+
+        yield STOP
+
+    def __len__(self):
+        if self._fake_codesize is not None:
+            return self._fake_codesize
+        return self._length_cache
+
+
+# ### section: sha3 preimage tracing
+def _stackitem_to_int(value):
+    if isinstance(value, tuple):
+        return to_int(value[1])  # how py-evm<=0.8.0b1 stores stuff on stack
+    else:
+        return to_int(value)
+
+
+def _stackitem_to_bytes(value):
+    if isinstance(value, tuple):
+        return to_bytes(value[1])  # how py-evm<=0.8.0b1 stores stuff on stack
+    else:
+        return to_bytes(value)
+
+
+class Sha3PreimageTracer:
+    mnemonic = "SHA3"
+
+    # trace preimages of sha3
+
+    def __init__(self, sha3_op, evm):
+        self.evm = evm
+        self.sha3 = sha3_op
+
+    def __call__(self, computation):
+        size, offset = [_stackitem_to_int(x) for x in computation._stack.values[-2:]]
+
+        # dispatch into py-evm
+        self.sha3(computation)
+
+        if size != 64:
+            return
+
+        preimage = computation._memory.read_bytes(offset, size)
+
+        image = _stackitem_to_bytes(computation._stack.values[-1])
+
+        self.evm._trace_sha3_preimage(preimage, image)
+
+
+class SstoreTracer:
+    mnemonic = "SSTORE"
+
+    def __init__(self, sstore_op, evm):
+        self.evm = evm
+        self.sstore = sstore_op
+
+    def __call__(self, computation):
+        value, slot = [_stackitem_to_int(t) for t in computation._stack.values[-2:]]
+        account = computation.msg.storage_address
+
+        self.evm._trace_sstore(account, slot)
+
+        # dispatch into py-evm
+        self.sstore(computation)
+
+
+# ### End section: sha3 tracing
+
+
+# py-evm uses class instantiaters which need to be classes
+# instead of like factories or other easier to use architectures -
+# `titanoboa_computation` is a class which can be constructed dynamically
+class titanoboa_computation:
+    _gas_meter_class = GasMeter
+
+    def __init__(self, *args, **kwargs):
+        # super() hardcodes CodeStream into the ctor
+        # so we have to override it here
+        super().__init__(*args, **kwargs)
+
+        self.code = TracingCodeStream(
+            self.code._raw_code_bytes,
+            fake_codesize=getattr(self.msg, "_fake_codesize", None),
+            start_pc=getattr(self.msg, "_start_pc", 0),
+        )
+        global _precompiles
+        # copy so as not to mess with class state
+        self._precompiles = self._precompiles.copy()
+        self._precompiles.update(_precompiles)
+
+        global _opcode_overrides
+        # copy so as not to mess with class state
+        self.opcodes = self.opcodes.copy()
+        self.opcodes.update(_opcode_overrides)
+
+        self._gas_meter = self._gas_meter_class(self.msg.gas)
+        if hasattr(self._gas_meter, "_set_code"):
+            self._gas_meter._set_code(self.code)
+
+        self._child_pcs = []
+        self._contract_repr_before_revert = None
+
+    def add_child_computation(self, child_computation):
+        super().add_child_computation(child_computation)
+        # track PCs of child calls for profiling purposes
+        self._child_pcs.append(self.code.program_counter)
+
+    # hijack creations to automatically generate blueprints
+    @classmethod
+    def apply_create_message(cls, state, msg, tx_ctx):
+        computation = super().apply_create_message(state, msg, tx_ctx)
+
+        bytecode = msg.code
+        # cf. eth/vm/logic/system/Create* opcodes
+        contract_address = msg.storage_address
+
+        if is_eip1167_contract(bytecode):
+            contract_address = extract_eip1167_address(bytecode)
+            bytecode = cls.env.vm.state.get_code(contract_address)
+
+        if bytecode in cls.env._code_registry:
+            target = cls.env._code_registry[bytecode].deployer.at(contract_address)
+            target.created_from = Address(msg.sender)
+            cls.env.register_contract(contract_address, target)
+
+        return computation
+
+    @classmethod
+    def apply_computation(cls, state, msg, tx_ctx):
+        addr = msg.code_address
+        contract = cls.env._lookup_contract_fast(addr) if addr else None
+
+        def finalize(c):
+            if c.is_error:
+                # After the computation is applied with an error the state is
+                # reverted. Before the revert, save the contract repr for the
+                # error message
+                c._contract_repr_before_revert = repr(contract)
+            return c
+
+        if contract is None or not cls.env._fast_mode_enabled:
+            # print("SLOW MODE")
+            computation = super().apply_computation(state, msg, tx_ctx)
+            return finalize(computation)
+
+        with cls(state, msg, tx_ctx) as computation:
+            try:
+                if getattr(msg, "_ir_executor", None) is not None:
+                    # print("MSG HAS IR EXECUTOR")
+                    # this happens when bytecode is overridden, e.g.
+                    # for injected functions. note ir_executor is (correctly)
+                    # used for the outer computation only because on subcalls
+                    # a clean message is constructed for the child computation
+                    msg._ir_executor.exec(computation)
+                else:
+                    # print("REGULAR FAST MODE")
+                    contract.ir_executor.exec(computation)
+            except Halt:
+                pass
+
+        # return computation outside of with block; computation.__exit__
+        # swallows exceptions (including Revert).
+        return finalize(computation)
+
+
+# Message object with extra attrs we can use to thread things through
+# the execution context.
+class FakeMessage(Message):
+    def __init__(
+        self,
+        *args,
+        ir_executor=None,
+        fake_codesize=None,
+        start_pc=0,
+        contract=None,
+        **kwargs,
+    ):
+        super().__init__(*args, **kwargs)
+        self._ir_executor = ir_executor
+        self._fake_codesize = fake_codesize
+        self._start_pc = start_pc
+        self._contract = contract
+
+
+class PyEVM:
+    def __init__(self, env, fast_mode_enabled: bool):
+        self.chain = _make_chain()
+        self.sha3_trace = {}
+        self.sstore_trace = {}
+        self._init_vm(env, reset_traces=True, fast_mode_enabled=fast_mode_enabled)
+
+    def _init_vm(self, env, reset_traces: bool, fast_mode_enabled: bool):
+        self.vm = self.chain.get_vm()
+
+        self.vm.patch = VMPatcher(self.vm)
+
+        c = type(
+            "TitanoboaComputation",
+            (titanoboa_computation, self.vm.state.computation_class),
+            {"env": env},
+        )
+
+        if fast_mode_enabled:
+            patch_pyevm_state_object(self.vm.state)
+
+        self.vm.state.computation_class = c
+
+        # we usually want to reset the trace data structures
+        # but sometimes don't, give caller the option.
+        if reset_traces:
+            self.sha3_trace = {}
+            self.sstore_trace = {}
+
+        # patch in tracing opcodes
+        c.opcodes[0x20] = Sha3PreimageTracer(c.opcodes[0x20], self)
+        c.opcodes[0x55] = SstoreTracer(c.opcodes[0x55], self)
+
+    def _trace_sha3_preimage(self, preimage, image):
+        self.sha3_trace[image] = preimage
+
+    def _trace_sstore(self, account, slot):
+        self.sstore_trace.setdefault(account, set())
+        # we don't want to deal with snapshots/commits/reverts, so just
+        # register that the slot was touched and downstream can filter
+        # zero entries.
+        self.sstore_trace[account].add(slot)
+
+    def enable_fast_mode(self, flag: bool = True):
+        if flag:
+            patch_pyevm_state_object(self.vm.state)
+        else:
+            unpatch_pyevm_state_object(self.vm.state)
+
+    def fork(self, url, reset_traces: bool, fast_mode_enabled: bool, **kwargs):
+        kwargs["url"] = url
+        AccountDBFork._rpc_init_kwargs = kwargs
+        self.vm.__class__._state_class.account_db_class = AccountDBFork
+        self._init_vm(reset_traces, fast_mode_enabled)
+        block_info = self.vm.state._account_db._block_info
+
+        self.vm.patch.timestamp = int(block_info["timestamp"], 16)
+        self.vm.patch.block_number = int(block_info["number"], 16)
+        # TODO patch the other stuff
+
+    def get_gas_meter_class(self):
+        return self.vm.state.computation_class._gas_meter_class
+
+    def set_gas_meter_class(self, cls: type):
+        self.vm.state.computation_class._gas_meter_class = cls
+
+    def get_balance(self, address: Address):
+        return self.vm.state.get_balance(address.canonical_address)
+
+    def set_balance(self, address: Address, value):
+        self.vm.state.set_balance(address.canonical_address, value)
+
+    def get_code(self, address: Address) -> bytes:
+        return self.vm.state.get_code(address.canonical_address)
+
+    def get_gas_limit(self):
+        return self.vm.state.gas_limit
+
+    # advanced: reset warm/cold counters for addresses and storage
+    def reset_access_counters(self):
+        self.vm.state._account_db._reset_access_counters()
+
+    def snapshot(self) -> (Hash32, JournalDBCheckpoint):
+        return self.vm.state.snapshot()
+
+    def anchor(self):
+        return self.vm.patch.anchor()
+
+    def revert(self, snapshot_id: (Hash32, JournalDBCheckpoint)) -> None:
+        self.vm.state.revert(snapshot_id)
+
+    def generate_contract_address(self, sender: Address):
+        nonce = self.vm.state.get_nonce(sender.canonical_address)
+        self.vm.state.increment_nonce(sender.canonical_address)
+        return Address(generate_contract_address(sender.canonical_address, nonce))
+
+    def deploy_code(
+        self,
+        sender: Address,
+        origin: Address,
+        target_address: Address,
+        prefetch_state: bool,
+        gas: Optional[int],
+        gas_price: int,
+        value: int,
+        bytecode: bytes,
+    ):
+        if gas is None:
+            gas = self.vm.state.gas_limit
+
+        msg = Message(
+            to=constants.CREATE_CONTRACT_ADDRESS,  # i.e., b""
+            sender=sender.canonical_address,
+            gas=gas,
+            value=value,
+            code=bytecode,
+            create_address=target_address.canonical_address,
+            data=b"",
+        )
+
+        if prefetch_state:
+            self.vm.state._account_db.try_prefetch_state(msg)
+
+        tx_ctx = BaseTransactionContext(
+            origin=origin.canonical_address, gas_price=gas_price
+        )
+        return self.vm.state.computation_class.apply_create_message(
+            self.vm.state, msg, tx_ctx
+        )
+
+    def execute_code(
+        self,
+        sender: Address,
+        to: Address,
+        prefetch_state: bool,
+        gas: int,
+        gas_price: int,
+        value: int,
+        bytecode: bytes,
+        data: bytes,
+        is_static: bool,
+        fake_codesize: Optional[int],
+        start_pc: int,
+        ir_executor: Any,
+        contract: Any,
+    ):
+        msg = FakeMessage(
+            sender=sender.canonical_address,
+            to=to.canonical_address,
+            gas=gas,
+            value=value,
+            code=bytecode,  # type: ignore
+            data=data,
+            is_static=is_static,
+            fake_codesize=fake_codesize,
+            start_pc=start_pc,
+            ir_executor=ir_executor,
+            contract=contract,
+        )
+        if prefetch_state:
+            self.vm.state._account_db.try_prefetch_state(msg)
+
+        origin = sender.canonical_address  # XXX: consider making this parameterizable
+        tx_ctx = BaseTransactionContext(origin=origin, gas_price=gas_price)
+        return self.vm.state.computation_class.apply_message(self.vm.state, msg, tx_ctx)
+
+    @property
+    def block_id(self):
+        return self.vm.state._account_db._block_id
+
+    def time_travel(self, add_seconds: int, add_blocks: int):
+        self.vm.patch.timestamp += add_seconds
+        self.vm.patch.block_number += add_blocks
+
+    def get_account_db(self):
+        return self.vm.state._account_db
+
+
+GENESIS_PARAMS = {"difficulty": constants.GENESIS_DIFFICULTY, "gas_limit": int(1e8)}
+
+
+# TODO make fork configurable - ex. "latest", "frontier", "berlin"
+# TODO make genesis params+state configurable
+def _make_chain():
+    # TODO should we use MiningChain? is there a perf difference?
+    # TODO debug why `fork_at()` cannot accept 0 as block num
+    _Chain = chain.build(MainnetChain, chain.latest_mainnet_at(1))
+    return _Chain.from_genesis(AtomicDB(), GENESIS_PARAMS)
