+++ tests/unitary/test_blueprints.py
@@ -1,5 +1,3 @@
-from eth_utils import to_canonical_address
-
 import boa
 from boa.util.eip5202 import get_create2_address
 
@@ -25,9 +23,7 @@ def test_create2_address():
     child_contract_address = factory.create_child(blueprint.address, salt)
 
     # TODO: make a util function on boa.env to get code
-    blueprint_bytecode = boa.env.vm.state.get_code(
-        to_canonical_address(blueprint.address)
-    )
+    blueprint_bytecode = boa.env.evm.get_code(blueprint.address)
     assert child_contract_address == get_create2_address(
         blueprint_bytecode, factory.address, salt
     )
